+++ fabric/src/main/java/dev/muon/dynamic_difficulty/compat/jade/LevelComponentProvider.java
@@ -2,7 +2,7 @@ package dev.muon.dynamic_difficulty.compat.jade;
 
 import dev.muon.dynamic_difficulty.DynamicDifficulty;
 import dev.muon.dynamic_difficulty.api.LevelingAPI;
-import dev.muon.dynamic_difficulty.client.LevelPlateRenderer;
+import dev.muon.dynamic_difficulty.client.LevelPlateHandler;
 import dev.muon.dynamic_difficulty.config.Config;
 import net.minecraft.client.Minecraft;
 import net.minecraft.network.chat.Component;
@@ -38,7 +38,7 @@ public enum LevelComponentProvider implements IEntityComponentProvider {
             if (!showLevel) return;
             
             int entityLevel = LevelingAPI.getLevel(living);
-            int color = LevelPlateRenderer.getLevelColorRGB(Minecraft.getInstance().player, living);
+            int color = LevelPlateHandler.getLevelColorRGB(Minecraft.getInstance().player, living);
             
             tooltip.add(Component.translatable("jade.dynamic_difficulty.tooltip", entityLevel)
                     .withStyle(style -> style.withColor(color)));